# Common/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [{
            "url": "http://localhost:9889",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/expense/{id}": {
            "get": {
                "tags": ["expense-controller"],
                "operationId": "get_13",
                "parameters": [{
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "default response",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Expense"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/expense/list/{id}": {
            "get": {
                "tags": ["expense-controller"],
                "operationId": "list_8",
                "parameters": [{
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "default response",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Expense"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/expense/delete/{id}": {
            "post": {
                "tags": ["expense-controller"],
                "operationId": "deleteExpense",
                "parameters": [{
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "default response"
                    }
                }
            }
        },
        "/api/expense/create": {
            "post": {
                "tags": ["expense-controller"],
                "operationId": "createExpense",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Expense"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "default response",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Expense"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Expense": {
                "type": "object",
                "properties": {
                    "expense_id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string"
                    },
                    "category_id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "account_id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "amount": {
                        "type": "number",
                        "format": "double"
                    }
                }
            },
...
        }
    }
}